Tilemap Render Integration
- Moved non-test related code into map.rs for use in the main code base
- Added getter functions that were missing
- Hooked up new map rendering functionality
- Updated Camera render function to spawn at map center
- Implemented various Clippy recommended code optimizations
