+++ tests/tilemap.rs
@@ -5,9 +5,9 @@ use bevy::render::settings::WgpuSettings;
 use bevy::render::RenderPlugin;
 use bevy::sprite::SpritePlugin;
 
-use cucumber::{given, then, when, World};
+use helping_hand::visuals::map::*;
 
-use tiled::{Loader, Map};
+use cucumber::{given, then, when, World};
 
 use std::ffi::OsString;
 use std::path::PathBuf;
@@ -21,237 +21,6 @@ struct GameWorld {
 
     pub loaded_map: Tilemap,
     pub bevy_map: RenderedMap,
-}
-
-#[derive(Debug)]
-struct Tilemap {
-    tiled_tiles: Vec<Tile>,
-    grid_dimensions: MapGridDimensions,
-    px_dimensions: PxDimensions,
-}
-
-impl Tilemap {
-    pub fn new(map_location: PathBuf) -> Self {
-        let mut loader = Loader::new();
-        let tiled_map = loader.load_tmx_map(map_location).unwrap();
-        let num_layers = tiled_map.layers().len() as u32;
-
-        let px_dimensions = Self::get_map_in_px(&tiled_map);
-        let tiled_tiles = get_map_tiles(tiled_map);
-
-        let num_rows = get_num_rows_from_map(&tiled_tiles);
-        let num_columns = get_num_columns_from_map(&tiled_tiles);
-        let grid_dimensions = MapGridDimensions::new(num_rows, num_columns, num_layers);
-
-        Self {
-            tiled_tiles,
-            grid_dimensions,
-            px_dimensions,
-        }
-    }
-
-    pub fn get_tiles(&self) -> &Vec<Tile> {
-        &self.tiled_tiles
-    }
-
-    pub fn get_grid_dimensions(&self) -> &MapGridDimensions {
-        &self.grid_dimensions
-    }
-
-    pub fn get_px_dimensions(&self) -> &PxDimensions {
-        &self.px_dimensions
-    }
-
-    pub fn tiles_overlap(&self, first_tile_index: usize, second_tile_index: usize) -> bool {
-        let first_tile_px_position = &self.tiled_tiles[first_tile_index].px_cords;
-        let second_tile_px_position = &self.tiled_tiles[second_tile_index].px_cords;
-
-        if first_tile_px_position.px_x == second_tile_px_position.px_x
-            && first_tile_px_position.px_y == second_tile_px_position.px_y
-        {
-            true
-        } else {
-            false
-        }
-    }
-
-    fn get_map_in_px(tiled_map: &Map) -> PxDimensions {
-        let px_dimensions = PxDimensions::new(
-            tiled_map.width * &tiled_map.tile_width,
-            tiled_map.height * &tiled_map.tile_height,
-        );
-
-        px_dimensions
-    }
-}
-
-impl Default for Tilemap {
-    fn default() -> Self {
-        Self {
-            tiled_tiles: Vec::new(),
-            grid_dimensions: MapGridDimensions::new(0, 0, 0),
-            px_dimensions: PxDimensions::new(0, 0),
-        }
-    }
-}
-
-#[derive(Debug, PartialEq)]
-struct XyzCords {
-    px_x: usize,
-    px_y: usize,
-    px_z: usize,
-}
-
-impl XyzCords {
-    pub fn new(px_x: u32, px_y: u32, px_z: usize) -> Self {
-        XyzCords {
-            px_x: px_x as usize,
-            px_y: px_y as usize,
-            px_z,
-        }
-    }
-}
-
-#[derive(Debug, PartialEq)]
-struct PxDimensions {
-    px_width: usize,
-    px_height: usize,
-}
-
-impl PxDimensions {
-    pub fn new(px_width: u32, px_height: u32) -> Self {
-        PxDimensions {
-            px_width: px_width as usize,
-            px_height: px_height as usize,
-        }
-    }
-}
-
-#[derive(Debug, PartialEq)]
-struct MapGridDimensions {
-    rows: u32,
-    columns: u32,
-    layers: u32,
-}
-
-impl MapGridDimensions {
-    pub fn new(rows: u32, columns: u32, layers: u32) -> MapGridDimensions {
-        MapGridDimensions {
-            rows,
-            columns,
-            layers,
-        }
-    }
-
-    pub fn get_layers(&self) -> u32 {
-        self.layers
-    }
-}
-
-#[derive(Debug)]
-struct Tile {
-    tile_dimensions: PxDimensions,
-    px_cords: XyzCords,
-    tile_texture: Option<TileTexture>,
-    layer_number: usize,
-}
-
-impl Tile {
-    pub fn new(
-        tile_dimensions: PxDimensions,
-        px_cords: XyzCords,
-        tile_texture: Option<TileTexture>,
-        layer_number: usize,
-    ) -> Tile {
-        Tile {
-            tile_dimensions,
-            px_cords,
-            tile_texture,
-            layer_number,
-        }
-    }
-
-    pub fn get_tile_spritesheet_filename(&self) -> OsString {
-        let tiles_texture = self.tile_texture.as_ref().unwrap();
-        let spritesheet_name = tiles_texture.spritesheet.file_name().unwrap();
-
-        spritesheet_name.into()
-    }
-
-    pub fn get_spritesheet_dimensions(&self) -> &PxDimensions {
-        let tile_texture = self.tile_texture.as_ref().unwrap();
-        &tile_texture.spritesheet_dimensions
-    }
-
-    pub fn get_sprite_index(&self) -> usize {
-        self.tile_texture.as_ref().unwrap().sprite_index
-    }
-}
-
-#[derive(Debug)]
-struct TileTexture {
-    spritesheet: PathBuf,
-    sprite_index: usize,
-    spritesheet_dimensions: PxDimensions,
-}
-
-#[derive(Bundle)]
-struct RenderTile {
-    spritesheet_bundle: SpriteSheetBundle,
-}
-
-#[derive(Default)]
-struct RenderedMap {
-    bevy_tiles: Vec<Option<RenderTile>>,
-}
-
-impl Debug for RenderedMap {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("RenderedMap")
-            .field(
-                "bevy_tiles",
-                &format_args!(
-                    "{}",
-                    self.bevy_tiles.iter().filter(|tile| tile.is_some()).count()
-                ),
-            )
-            .finish()
-    }
-}
-
-impl RenderedMap {
-    pub fn new(
-        tilemap: &Tilemap,
-        asset_server: &AssetServer,
-        texture_atlas_assets: &mut Assets<TextureAtlasLayout>,
-    ) -> Self {
-        RenderedMap {
-            bevy_tiles: get_render_tile_bundles(&tilemap, &asset_server, texture_atlas_assets),
-        }
-    }
-
-    pub fn tiled_map_overlap(
-        &self,
-        tiled_map: &Tilemap,
-        tiled_tile_index: usize,
-        bevy_tile_index: usize,
-    ) -> bool {
-        let tiled_tile_px_position = &tiled_map.get_tiles()[tiled_tile_index].px_cords;
-        let bevy_tile_px_position = &self.bevy_tiles[bevy_tile_index]
-            .as_ref()
-            .unwrap()
-            .spritesheet_bundle
-            .transform
-            .translation;
-
-        if tiled_tile_px_position.px_x == bevy_tile_px_position.x as usize
-            && tiled_tile_px_position.px_y == bevy_tile_px_position.y as usize
-        {
-            true
-        } else {
-            false
-        }
-    }
 }
 
 impl GameWorld {
@@ -274,7 +43,7 @@ impl GameWorld {
             .into(),
             ..default()
         });
-        app.add_plugins(SpritePlugin::default());
+        app.add_plugins(SpritePlugin);
         app.add_plugins(ImagePlugin::default());
 
         Self {
@@ -285,165 +54,6 @@ impl GameWorld {
             assets_folder_path: absolute_assets_folder_path,
         }
     }
-}
-
-fn get_map_tiles(tiled_map: Map) -> Vec<Tile> {
-    let tile_width = tiled_map.tile_width;
-    let tile_height = tiled_map.tile_height;
-
-    let map_width = tiled_map.width;
-    let map_height = tiled_map.height;
-
-    let mut tiles = Vec::new();
-
-    for z in 0..tiled_map.layers().len() {
-        for y in 0..map_height {
-            for x in 0..map_width {
-                let tile = Tile {
-                    tile_dimensions: PxDimensions::new(tile_width, tile_height),
-                    px_cords: XyzCords::new(x * tile_width, y * tile_height, z),
-                    layer_number: z,
-                    tile_texture: get_tile_texture(&tiled_map, x, y, z),
-                };
-                tiles.push(tile);
-            }
-        }
-    }
-
-    tiles
-}
-
-/// Returns a list of RenderTileBundles to be spawned by Bevy for the given list of tiles.
-fn get_render_tile_bundles(
-    tilemap: &Tilemap,
-    asset_server: &AssetServer,
-    texture_atlas_assets: &mut Assets<TextureAtlasLayout>,
-) -> Vec<Option<RenderTile>> {
-    let mut render_tile_bundles = Vec::new();
-
-    let tiles = tilemap.get_tiles();
-
-    for tile in tiles {
-        //Tiles without a texture don't need to be rendered
-        if tile.tile_texture.is_none() {
-            render_tile_bundles.push(None);
-            continue;
-        }
-
-        //We have to trim our path from being absolute to having root at assets
-        let bevy_path = to_bevy_path(&tile.tile_texture.as_ref().unwrap().spritesheet);
-        let texture = asset_server.load(bevy_path);
-
-        //Getting Spritesheet Dimensions
-        let sprite_sheet_column_count = (tile
-            .tile_texture
-            .as_ref()
-            .unwrap()
-            .spritesheet_dimensions
-            .px_width
-            / tile.tile_dimensions.px_width) as usize;
-        let sprite_sheet_row_count = (tile
-            .tile_texture
-            .as_ref()
-            .unwrap()
-            .spritesheet_dimensions
-            .px_height
-            / tile.tile_dimensions.px_height) as usize;
-
-        //This is how the sprite sheet should be cut when creating sprites to render
-        let sheet_layout = TextureAtlasLayout::from_grid(
-            Vec2::new(
-                tile.tile_dimensions.px_width as f32,
-                tile.tile_dimensions.px_height as f32,
-            ),
-            sprite_sheet_column_count,
-            sprite_sheet_row_count,
-            None,
-            None,
-        );
-
-        // Conversion to Bevy specific formatting happens right here
-        // Our:RenderTileBundle -> Bevy's:SpriteSheetBundle
-        let render_tile = Some(RenderTile {
-            spritesheet_bundle: SpriteSheetBundle {
-                transform: Transform::from_xyz(
-                    tile.px_cords.px_x as f32,
-                    //y-axis flip because Bevy is Y-Up while Tiled is Y-Down
-                    flip_y_axis(
-                        tilemap.get_px_dimensions().px_height,
-                        tile.px_cords.px_y as f32,
-                        tile.tile_dimensions.px_height,
-                    ),
-                    tile.px_cords.px_z as f32,
-                ),
-                texture,
-                atlas: TextureAtlas {
-                    layout: texture_atlas_assets.add(sheet_layout),
-                    index: tile.tile_texture.as_ref().unwrap().sprite_index,
-                },
-                ..Default::default()
-            },
-        });
-        render_tile_bundles.push(render_tile);
-    }
-    render_tile_bundles
-}
-
-fn get_tile_texture(
-    tiled_map: &Map,
-    x_grid_cord: u32,
-    y_grid_cord: u32,
-    layer_num: usize,
-) -> Option<TileTexture> {
-    let tile_layer = tiled_map
-        .get_layer(layer_num)
-        .unwrap()
-        .as_tile_layer()
-        .unwrap();
-
-    if let Some(tile) = tile_layer.get_tile(x_grid_cord as i32, y_grid_cord as i32) {
-        tile_layer
-            .get_tile(x_grid_cord as i32, y_grid_cord as i32)
-            .unwrap();
-
-        let sprite_index = tile.id() as usize;
-        let spritesheet = tile.get_tileset().image.clone().unwrap().source;
-        let spritesheet_px_width = tile.get_tileset().image.as_ref().unwrap().width as u32;
-        let spritesheet_px_height = tile.get_tileset().image.as_ref().unwrap().height as u32;
-
-        Some(TileTexture {
-            sprite_index,
-            spritesheet,
-            spritesheet_dimensions: PxDimensions::new(spritesheet_px_width, spritesheet_px_height),
-        })
-    } else {
-        None
-    }
-}
-
-fn to_bevy_path(tiled_path: &PathBuf) -> PathBuf {
-    let mut trimmed_path = PathBuf::new();
-    let mut path_element_stack = Vec::new();
-
-    for path_element in tiled_path.iter().rev() {
-        if path_element == "assets" {
-            break;
-        }
-
-        path_element_stack.push(path_element);
-    }
-
-    while let Some(path_element) = path_element_stack.pop() {
-        trimmed_path.push(path_element);
-    }
-
-    return trimmed_path;
-}
-
-fn flip_y_axis(map_height: usize, tile_y: f32, tile_height: usize) -> f32 {
-    let flipped_y = map_height as f32 - tile_y - tile_height as f32;
-
-    return flipped_y;
 }
 
 // Returns a Path to the specified Tiled Map
@@ -461,36 +71,6 @@ fn get_tiled_map_location(map_name: String) -> PathBuf {
     tiled_map_path.push(map_name);
 
     tiled_map_path
-}
-
-fn get_num_columns_from_map(tiles: &[Tile]) -> u32 {
-    let mut highest_x = 0;
-    let tile_width = tiles[0].tile_dimensions.px_width;
-
-    for tile in tiles {
-        if tile.px_cords.px_x >= highest_x {
-            highest_x = tile.px_cords.px_x;
-        } else {
-            break;
-        }
-    }
-
-    ((highest_x / tile_width) + 1) as u32
-}
-
-fn get_num_rows_from_map(tiles: &[Tile]) -> u32 {
-    let mut highest_y = 0;
-    let tile_height = tiles[0].tile_dimensions.px_height;
-
-    for tile in tiles {
-        if tile.px_cords.px_y >= highest_y {
-            highest_y = tile.px_cords.px_y;
-        } else {
-            break;
-        }
-    }
-
-    ((highest_y / tile_height) + 1) as u32
 }
 
 //////////////TEST FUNCTIONS//////////////
@@ -532,7 +112,7 @@ fn tiled_map_to_bevy_tiles(world: &mut GameWorld) {
     let asset_server = world.app.world.resource::<AssetServer>().clone();
     let mut texture_atlas_layout = world.app.world.resource_mut::<Assets<TextureAtlasLayout>>();
 
-    let rendered_bevy_map = RenderedMap::new(&tilemap, &asset_server, &mut texture_atlas_layout);
+    let rendered_bevy_map = RenderedMap::new(tilemap, &asset_server, &mut texture_atlas_layout);
     world.bevy_map = rendered_bevy_map;
 }
 
@@ -620,7 +200,7 @@ fn verify_tile_contains_image(world: &mut GameWorld, tile_num: String) {
         .parse::<usize>()
         .expect("verify_tile_contains_image: tile_num is not a number?");
 
-    let tile_image = &world.loaded_map.tiled_tiles[tile_index - 1].tile_texture;
+    let tile_image = &world.loaded_map.get_tiles()[tile_index - 1].get_tile_texture();
 
     assert!(tile_image.is_some());
 }
@@ -631,7 +211,7 @@ fn verify_tile_image_is_empty(world: &mut GameWorld, tile_num: String) {
         .parse::<usize>()
         .expect("verify_tile_contains_image: tile_num is not a number?");
 
-    let tile_image = &world.loaded_map.get_tiles()[tile_index - 1].tile_texture;
+    let tile_image = &world.loaded_map.get_tiles()[tile_index - 1].get_tile_texture();
     assert!(tile_image.is_none());
 }
 
@@ -688,7 +268,7 @@ fn verify_render_tile_is_some(world: &mut GameWorld, tile_num: String) {
         .parse::<usize>()
         .expect("verify_render_tile_is_some: tile_num is not a number?");
 
-    let rendered_tile = &world.bevy_map.bevy_tiles[tile_index - 1];
+    let rendered_tile = &world.bevy_map.get_bevy_tiles()[tile_index - 1];
     assert!(rendered_tile.is_some());
 }
 
@@ -698,7 +278,7 @@ fn verify_render_tile_is_none(world: &mut GameWorld, tile_num: String) {
         .parse::<usize>()
         .expect("verify_render_tile_is_non: tile_num is not a number?");
 
-    let rendered_tile = &world.bevy_map.bevy_tiles[tile_index - 1];
+    let rendered_tile = &world.bevy_map.get_bevy_tiles()[tile_index - 1];
     assert!(rendered_tile.is_none())
 }
